third_party/WebKit/LayoutTests/media/media-can-play-type-expected.txt

Test HTMLMediaElement canPlayType() method.

TEST(video.canPlayType()) THROWS("TypeError: Failed to execute 'canPlayType' on 'HTMLMediaElement': 1 argument required, but only 0 present.") OK
EXPECTED (video.canPlayType('video/') == '') OK
EXPECTED (video.canPlayType('video/blahblah') == '') OK
EXPECTED (video.canPlayType('video/blahblah; codecs="blah, , blah"') == '') OK
EXPECTED (video.canPlayType('video/blahblah; codecs=blah') == '') OK
EXPECTED (video.canPlayType('video/blahblah; codecs="blah"') == '') OK
EXPECTED (video.canPlayType('video/blahblah; codecs="badcontent') == '') OK
EXPECTED (video.canPlayType('video/blahblah; codecs=badcontent"') == '') OK
EXPECTED (video.canPlayType('video/blahblah; codecs="badcontent"') == '') OK
END OF TEST
